feat: deep audit + coverage push + hotSpring trace stabilization (Iter 64)
- Clean up hotSpring's incomplete trace module (missing trace.rs, stale
  imports, trait methods, cross-primal references)
- Implement OpBRev and OpFRnd encoding for SM32/SM50 encoders
- Smart-refactor 5 files over 1000 LOC into directory modules:
  acr_boot (4462→12), coralctl (1649→6), socket (1434→3),
  mmu_oracle (1131→3), device (1030→2)
- Resolve 222 clippy warnings, 13 doc link issues
- Add ~500 new tests across all testable crates (3460→3912)
- Coverage: 62.9%→65.9% workspace, 81.5% non-hardware
  coralreef-core 96%, stubs 97.7%, compiler 82%
- Update root docs and debt report for Iter 64

Quality gates: fmt clean, clippy zero warnings, 3912 tests pass,
doc clean (2 pre-existing coral-driver warnings)

Made-with: Cursor

+++ crates/coral-reef/src/lib.rs
@@ -208,7 +208,7 @@ fn shader_model_for(target: GpuTarget) -> Result<Box<dyn codegen::ir::ShaderMode
         }
         GpuTarget::Amd(amd) => {
             let gfx = amd.gfx_major() * 10 + 3;
-            let wave = u8::from(amd.default_wave_size());
+            let wave = amd.default_wave_size();
             Ok(Box::new(
                 codegen::amd::shader_model::ShaderModelRdna2::new(gfx).with_wave_size(wave),
             ))
